Refactoring

+++ javafx/src/netflow/model/MNode.java
@@ -1,5 +1,5 @@
 package netflow.model;
 
 public class MNode {
-    public float flow = 0, capacity = 10;
+    public double flow = 0, capacity = 10;
 }
